Fix build

+++ source/native-plugins/midi-sequencer.cpp
@@ -253,10 +253,10 @@ protected:
             const double beatsPerMinute = fTimeInfo.bbt.valid ? fTimeInfo.bbt.beatsPerMinute : 120.0;
             const float  beatType       = fTimeInfo.bbt.valid ? fTimeInfo.bbt.beatType       : 4.0f;
 
-            const double ticksPerBeat   = 48.0;
-            const double ticksPerFrame  = ticksPerBeat / (60.0 / beatsPerMinute * getSampleRate());
-            const long double fullTicks = ticksPerFrame*static_cast<long double>(fTimeInfo.frame);
-            const      double fullBeats = fullTicks/ticksPerBeat;
+            const double ticksPerBeat  = 48.0;
+            const double ticksPerFrame = ticksPerBeat / (60.0 / beatsPerMinute * getSampleRate());
+            const double fullTicks     = static_cast<double>(ticksPerFrame*static_cast<long double>(fTimeInfo.frame));
+            const double fullBeats     = fullTicks/ticksPerBeat;
 
             const uint32_t tick = static_cast<uint32_t>(std::floor(std::fmod(fullTicks, ticksPerBeat)));
             const uint32_t beat = static_cast<uint32_t>(std::floor(std::fmod(fullBeats, beatsPerBar)));
